refactor(#38): 모달 파일 구조 생성 및 index.ts 엔트리화
